+++ eduVPN/ProviderService.swift
@@ -496,7 +496,7 @@ class ProviderService {
                         
                         let publicKey = instance["public_key"] as? String
                         
-                        return Provider(displayName: displayName, baseURL: baseURL, logoURL: logoURL, publicKey: publicKey, connectionType: connectionType, authorizationType: authorizationType)
+                        return Provider(displayName: displayName, baseURL: baseURL, logoURL: logoURL, publicKey: publicKey, username: nil, connectionType: connectionType, authorizationType: authorizationType)
                     }
                     
                     guard !providers.isEmpty else {
@@ -875,7 +875,7 @@ class ProviderService {
                     let importedConfigFileURL = try localConfigsDirectoryURL.appendingPathComponent(displayName).nextUnusedFileURL()
                     displayName = importedConfigFileURL.lastPathComponent
                     try config.write(to: importedConfigFileURL, atomically: true, encoding: .utf8)
-                    let provider = Provider(displayName: displayName, baseURL: importedConfigFileURL, logoURL: nil, publicKey: commonName, connectionType: .localConfig, authorizationType: .local)
+                    let provider = Provider(displayName: displayName, baseURL: importedConfigFileURL, logoURL: nil, publicKey: commonName, username: nil, connectionType: .localConfig, authorizationType: .local)
                     if self.storedProviders[.localConfig] != nil {
                         self.storedProviders[.localConfig]?.append(provider)
                     } else {
@@ -975,12 +975,12 @@ class ProviderService {
         }
     }
     
-    func saveCommonCertificate(_ name: String, for provider: Provider) {
+    func saveCommonCertificate(_ name: String?, username: String?, for provider: Provider) {
         guard provider.connectionType == .localConfig else {
             return
         }
         
-        guard provider.publicKey != name else {
+        guard provider.publicKey != name, provider.username != username else {
             return
         }
         
@@ -989,10 +989,12 @@ class ProviderService {
         if let index = index {
             var storedProvider = providers.remove(at: index)
             storedProvider.publicKey = name
+            storedProvider.username = username
             providers.insert(storedProvider, at: index)
         } else {
             var newProvider = provider
             newProvider.publicKey = name
+            newProvider.username = username
             providers.append(newProvider)
         }